refactor(vscode-extension): centralize agent call logic with dispatch utility
- Introduced a new utility function, dispatchAgentCall, to handle IDE-specific agent calls for Cursor and Windsurf.
- Removed redundant agent call logic from activation and setup-toolbar files, streamlining the codebase.
- Enhanced error messaging for unsupported IDEs.
