+++ apps/vscode-extension/src/activation/activate.ts
@@ -1,15 +1,14 @@
 import * as vscode from 'vscode';
+import { dispatchAgentCall } from 'src/utils/dispatch-agent-call';
 import { startServer, stopServer } from '../http-server/server';
 import { updateCursorMcpConfig } from './register-mcp-server';
 import { findAvailablePort } from '../utils/find-available-port';
-import { callCursorAgent } from '../utils/call-cursor-agent';
 import {
   getExtensionBridge,
   DEFAULT_PORT,
 } from '@stagewise/extension-toolbar-srpc-contract';
 import { setupToolbar } from './setup-toolbar';
 import { getCurrentIDE } from 'src/utils/get-current-ide';
-import { callWindsurfAgent } from 'src/utils/call-windsurf-agent';
 
 // Diagnostic collection specifically for our fake prompt
 const fakeDiagCollection =
@@ -43,7 +42,7 @@ export async function activate(context: vscode.ExtensionContext) {
 
     bridge.register({
       triggerAgentPrompt: async (request, sendUpdate) => {
-        await determineAndCallAgent(request.prompt);
+        await dispatchAgentCall(request.prompt);
         sendUpdate.sendUpdate({ updateText: 'Called the agent' });
 
         return { result: { success: true } };
@@ -64,19 +63,4 @@ export async function activate(context: vscode.ExtensionContext) {
 
 export async function deactivate() {
   await stopServer();
-}
-
-export async function determineAndCallAgent(prompt: string) {
-  const ide = getCurrentIDE();
-  switch (ide) {
-    case 'CURSOR':
-      return await callCursorAgent(prompt);
-    case 'WINDSURF':
-      return await callWindsurfAgent(prompt);
-    case 'VSCODE':
-    case 'UNKNOWN':
-      vscode.window.showErrorMessage(
-        'Failed to call agent: IDE is not supported',
-      );
-  }
 }
